FileService: a way to work with file at frontend
The usecase of this code is something like this: let image that user is making a
protonmail or any another application where content inside frontend (=browser
storage) some secret user's data that shouldn't be exposed to backend. Secret
key for example or anything like that.

This code allows to developer to convert any `Seq[Array[Byte]]` from
frontend to URL as simple call `FileService.createURL`, asynchronously
convert any uploaded `File` to `Future[Array[Byte]]` as
`FileService.asBytesArray`, or to `Array[Byte]` via
`FileService.asSyncBytesArray` inside worker.

Unfortunately scalatags doesn't support `download` attribute and I need
to make it by hand. I've opened a PR[^1] to introduce it, but it might be a
while until it is included to release.

`FileService.asSyncBytesArray` is using `FileReaderSync` that is also
missed inside scala-js-dom. I've opened a PR[^2] but it might be a
while. Also, this draft API but it is supported by majority of modern
browsers[^3].

[^1]: https://github.com/lihaoyi/scalatags/pull/212

[^2]: https://github.com/scala-js/scala-js-dom/pull/424

[^3]: https://caniuse.com/?search=FileReaderSync